+++ Sources/VComponents/Components/Modals/Menu/VMenuRow.swift
@@ -7,88 +7,127 @@
 
 import SwiftUI
 
-// MARK: - V Menu Row
-/// Model that represents menu row, such as title, title with icons, or sub-menu.
-public struct VMenuRow {
+// MARK: - V Menu Section
+/// Single row in `VMenu`.
+public protocol VMenuRow: VMenuRowConvertible {
+    /// Row body type.
+    typealias Body = AnyView
+    
+    /// Row body.
+    var body: Body { get }
+}
+
+extension VMenuRow {
+    public func toRows() -> [any VMenuRow] { [self] }
+}
+
+// MARK: - V Menu Title Row
+/// SIngle row in `VMenu` with title.
+public struct VMenuTitleRow: VMenuRow {
     // MARK: Properties
-    let _menuRow: _VMenuRow
+    let action: () -> Void
+    let title: String
     
     // MARK: Initializers
-    private init(
-        menuRow: _VMenuRow
-    ) {
-        self._menuRow = menuRow
-    }
-    
-    /// Row with title.
-    public static func title(
+    /// Initializes `VMenuTitleRow` with action and title.
+    public init(
         action: @escaping () -> Void,
         title: String
-    ) -> Self {
-        .init(menuRow: .title(
-            action: action,
-            title: title
-        ))
+    ) {
+        self.action = action
+        self.title = title
     }
     
-    /// Row with title and icon from assets catalog.
-    public static func titleIcon(
+    // MARK: Row Section
+    public var body: AnyView {
+        .init(
+            Button(title, action: action)
+        )
+    }
+}
+
+// MARK: - V Menu Title Icon Row
+/// SIngle row in `VMenu` with title and icon.
+public struct VMenuTitleIconRow: VMenuRow {
+    // MARK: Properties
+    let action: () -> Void
+    let title: String
+    let icon: Image
+    
+    // MARK: Initializers
+    /// Initializes `VMenuTitleIconRow` with action, title, and icon.
+    public init(
+        action: @escaping () -> Void,
+        title: String,
+        icon: Image
+    ) {
+        self.action = action
+        self.title = title
+        self.icon = icon
+    }
+    
+    /// Initializes `VMenuTitleIconRow` with action, title, and asset icon name.
+    public init(
         action: @escaping () -> Void,
         title: String,
         assetIcon: String,
         bundle: Bundle? = nil
-    ) -> Self {
-        .init(menuRow: .titleAssetIcon(
-            action: action,
-            title: title,
-            icon: assetIcon,
-            bundle: bundle
-        ))
+    ) {
+        self.action = action
+        self.title = title
+        self.icon = .init(assetIcon, bundle: bundle)
     }
     
-    /// Row with title and icon.
-    public static func titleIcon(
-        action: @escaping () -> Void,
-        title: String,
-        icon: Image
-    ) -> Self {
-        .init(menuRow: .titleIcon(
-            action: action,
-            title: title,
-            icon: icon
-        ))
-    }
-    
-    /// Row with title and sytem icon.
-    public static func titleIcon(
+    /// Initializes `VMenuTitleIconRow` with action, title, and system icon name.
+    public init(
         action: @escaping () -> Void,
         title: String,
         systemIcon: String
-    ) -> Self {
-        .init(menuRow: .titleSystemIcon(
-            action: action,
-            title: title,
-            icon: systemIcon
-        ))
+    ) {
+        self.action = action
+        self.title = title
+        self.icon = .init(systemName: systemIcon)
     }
     
-    /// Row that expands to sub-menu.
-    public static func menu(
-        title: String,
-        rows: [VMenuRow]
-    ) -> Self {
-        .init(menuRow: .menu(
-            title: title,
-            rows: rows
-        ))
+    // MARK: Row Section
+    public var body: AnyView {
+        .init(
+            Button(action: action, label: {
+                Text(title)
+                icon
+            })
+        )
     }
 }
 
-// MARK: - _ V Menu Row
-enum _VMenuRow {
-    case title(action: () -> Void, title: String)
-    case titleAssetIcon(action: () -> Void, title: String, icon: String, bundle: Bundle?)
-    case titleIcon(action: () -> Void, title: String, icon: Image)
-    case titleSystemIcon(action: () -> Void, title: String, icon: String)
-    case menu(title: String, rows: [VMenuRow])
+// MARK: - V Menu Sub Menu Row
+/// SIngle row in `VMenu` with submenu.
+public struct VMenuSubMenuRow: VMenuRow {
+    // MARK: Properties
+    let title: String
+    let primaryAction: (() -> Void)?
+    private let sections: () -> [any VMenuSection]
+    
+    // MARK: Initializers
+    /// Initializes `VMenuPickerSection` title with sections.
+    public init(
+        title: String,
+        primaryAction: (() -> Void)? = nil,
+        @VMenuSectionBuilder sections: @escaping () -> [any VMenuSection]
+    ) {
+        self.title = title
+        self.primaryAction = primaryAction
+        self.sections = sections
+    }
+    
+    // MARK: Row Section
+    public var body: AnyView {
+        .init(
+            VMenu(
+                primaryAction: primaryAction,
+                title: title,
+                sections: sections
+            )
+        )
+    }
 }